requirements.txt
agent-protocol==1.0.2
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1734244716962/work
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
banks==2.1.0
beautifulsoup4==4.13.3
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1740094013202/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560564262/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
duckduckgo_search==7.5.5
fastapi==0.100.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
filetype==1.2.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1741403990995/work
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1735929639977/work
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1733462545272/work
greenlet==3.1.1
griffe==1.7.1
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.7
httpx==0.28.1
hypercorn==0.14.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
iniconfig==2.1.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1733472696581/work
jsonschema-specifications @ file:///tmp/tmpk0f344m9/src
langchain==0.3.21
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
llama-cloud==0.1.17
llama-index==0.10.68
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.12.27
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
lxml==5.3.1
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
marshmallow==3.26.1
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict==6.2.0
mypy-extensions==1.0.0
narwhals @ file:///home/conda/feedstock_root/build_artifacts/narwhals_1742841036354/work
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
nltk==3.9.1
numpy==1.26.4
ollama==0.4.7
openai==1.69.0
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1741963795194/work
orjson==3.10.16
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1726878443020/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929690690/work
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
pluggy==1.5.0
primp==0.14.0
priority==2.0.0
propcache==0.3.1
protobuf @ file:///home/conda/feedstock_root/build_artifacts/protobuf_1741124588522/work/bazel-bin/python/dist/protobuf-5.29.3-cp311-abi3-linux_x86_64.whl#sha256=1697af3d213954996732a0961863db588821a9b6641aed8358e87dc56d35ad70
pyarrow==19.0.1
pybind11 @ file:///D:/bld/pybind11-split_1730237329882/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1730237328107/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.1
pydantic_core==2.33.0
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
pypdf==4.3.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest==7.4.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.1.0
python-multipart==0.0.6
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-toolbelt==1.0.0
rpds-py @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rpds-py_1743037725/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.27.0
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1743193014327/work
striprtf==0.0.26
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1736248176451/work
tenacity==8.5.0
tiktoken==0.9.0
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
torch @ file:///home/conda/feedstock_root/build_artifacts/libtorch_1742916229386/work
torchaudio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_torchaudio_1742953096/work
torchvision @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_torchvision_1742619225/work
torchvision-extra-decoders @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_torchvision-extra-decoders_1739805887/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615904614/work
tqdm==4.67.1
trio==0.29.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1743201626/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3==1.26.20
watchdog @ file:///home/conda/feedstock_root/build_artifacts/watchdog_1730492882558/work
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
